Updated move function
Updated move function to include more randomized sprite movement (instead of the sprite just bouncing around the enclosure in straight lines). Included code to flip the sprite if the x-value of the speed variable changes from positive to negative, or vice versa. Added to the if statements that flip the sprite if it hits the enclosure walls so that the sprite doesn't get stuck due to flipping from the previous if statements.